Renamed and generalized package.
