+++ edx_rest_api_client/exceptions.py
