+++ edx_rest_api_client/tests/test_client.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 import ddt
 import mock
 
-from ecommerce_api_client.client import EcommerceApiClient
+from edx_rest_api_client.client import EdxRestApiClient
 
 
 URL = 'http://example.com/api/v2'
@@ -17,7 +17,7 @@ ACCESS_TOKEN = 'abc123'
 
 
 @ddt.ddt
-class EcommerceApiClientTests(TestCase):
+class EdxRestApiClientTests(TestCase):
     """ Tests for the E-Commerce API client. """
 
     @ddt.data(
@@ -28,7 +28,7 @@ class EcommerceApiClientTests(TestCase):
     )
     def test_valid_configuration(self, kwargs):
         """ The constructor should return successfully if all arguments are valid. """
-        EcommerceApiClient(**kwargs)
+        EdxRestApiClient(**kwargs)
 
     @ddt.data(
         {'url': None, 'signing_key': SIGNING_KEY, 'username': USERNAME},
@@ -38,17 +38,17 @@ class EcommerceApiClientTests(TestCase):
     )
     def test_invalid_configuration(self, kwargs):
         """ If the constructor arguments are invalid, an InvalidConfigurationError should be raised. """
-        self.assertRaises(ValueError, EcommerceApiClient, **kwargs)
+        self.assertRaises(ValueError, EdxRestApiClient, **kwargs)
 
-    @mock.patch('ecommerce_api_client.auth.JwtAuth.__init__', return_value=None)
+    @mock.patch('edx_rest_api_client.auth.JwtAuth.__init__', return_value=None)
     def test_tracking_context(self, mock_auth):
         """ Ensure the tracking context is included with API requests if specified. """
-        EcommerceApiClient(URL, SIGNING_KEY, USERNAME, FULL_NAME, EMAIL, tracking_context=TRACKING_CONTEXT)
+        EdxRestApiClient(URL, SIGNING_KEY, USERNAME, FULL_NAME, EMAIL, tracking_context=TRACKING_CONTEXT)
         self.assertIn(TRACKING_CONTEXT, mock_auth.call_args[1].values())
 
     def test_oauth2(self):
         """ Ensure OAuth2 authentication is used when an access token is supplied to the constructor. """
 
-        with mock.patch('ecommerce_api_client.auth.BearerAuth.__init__', return_value=None) as mock_auth:
-            EcommerceApiClient(URL, oauth_access_token=ACCESS_TOKEN)
+        with mock.patch('edx_rest_api_client.auth.BearerAuth.__init__', return_value=None) as mock_auth:
+            EdxRestApiClient(URL, oauth_access_token=ACCESS_TOKEN)
             mock_auth.assert_called_with(ACCESS_TOKEN)
